New Event/Listening System and new Window System for Module Settings. Known Problems:  - Auto Totem doesn't work  - some Window need an overhaul  - more Module need more Windows

+++ src/main/java/com/noadsch12/render/ui/keystrokes/KeystrokesConfig.java
@@ -19,7 +19,7 @@ package com.noadsch12.render.ui.keystrokes;
 
 public class KeystrokesConfig {
     public static float x = 20, y = 20;
-    public static float scale = 1.0f;
+    public static double scale = 1.0f;
     public static float opacity = 0.8f;
     public static int color = 0xFFFFFFFF; // White
